Thanks to @johnmai-dev, we were able to make minor performance optimizations on Swift Jinja package, more information can be found here: https://github.com/johnmai-dev/Jinja/pull/24.

Benchmark Results:

| Phase              | Before (ms) | After (ms) | Change | % Improvement |
|--------------------|-------------|------------|--------|---------------|
| **Tokenize avg**   | 69.525      | 68.887     | -0.638 | ~0.9% faster  |
| **Parse avg**      | 0.498       | 0.457      | -0.041 | ~8.2% faster  |
| **Run avg**        | 0.116       | 0.099      | -0.017 | ~14.7% faster |
| **Template.render**| 0.176       | 0.142      | -0.034 | ~19.3% faster |